requirements.txt
Django==1.11.5
django-appconf==1.0.2
django-choices==1.6.0
django-cors-headers==2.1.0
django-extra-fields==0.9
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
olefile==0.44
Pillow==4.2.1
PyJWT==1.5.3
pytz==2017.2
